requirements.txt
fastapi>=0.111.0
uvicorn>=0.29.0
requests>=2.32.3
openai>=1.70.0
anthropic>=0.40.0
python-dotenv>=1.0.1
pydantic>=2.8.0
